+++ config/medium/FixTheSpaces/code/CGCode.js
@@ -1,0 +1,7 @@
+const original = readline();
+const words = readline();
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('original sentence with spaces');